backend/requirements.txt
fastapi
uvicorn[standard]
supabase
google-generativeai
python-dotenv
pydantic
python-jose[cryptography]
httpx
slowapi
ruff
